sample_project/requirements.txt
alembic==1.5.8
aniso8601==9.0.1
attrs==20.3.0
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
pdfminer.six==20200517
pdfplumber==0.5.27
Pillow==8.1.2
pycryptodome==3.10.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-pptx==0.6.18
pytz==2021.1
six==1.15.0
sortedcontainers==2.3.0
SQLAlchemy==1.4.2
Wand==0.6.6
Werkzeug==1.0.1
XlsxWriter==1.3.7
